+++ examples/storage_volume_templates.py
@@ -17,7 +17,6 @@
 
 from pprint import pprint
 from hpOneView.oneview_client import OneViewClient
-from hpOneView.exceptions import HPOneViewException
 from config_loader import try_load_from_file
 
 config = {
@@ -31,15 +30,13 @@ config = {
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
 
-oneview_client = OneViewClient(config)
-
 networks = "/rest/fcoe-networks/7f0f74a0-4957-47ac-81c1-f573aa6d83de"
 scope_uris = "/rest/scopes/63d1ca81-95b3-41f1-a1ee-f9e1bc2d635f"
 
 # Request body for create operation
 # Supported from API version >= 500
 options = {
-    "rootTemplateUri": "/rest/storage-volume-templates/5dbaf127-053b-4988-82fe-a80800eef1f3",
+    "rootTemplateUri": "/rest/storage-volume-templates/b8c4489e-4a19-4bfe-857c-aab8006478a7",
     "properties": {
         "name": {
             "title": "Volume name",
@@ -94,7 +91,7 @@ options = {
             "format": "x-uri-reference",
             "required": "true",
             "description": "A common provisioning group URI reference",
-            "default": "/rest/storage-pools/0DB2A6C7-04D3-4830-9229-A80800EEF1F1"
+            "default": "/rest/storage-pools/628C1EBD-5BA7-40F2-A856-A93C0143AC73"
         },
         "snapshotPool": {
             "meta": {
@@ -104,7 +101,7 @@ options = {
             "type": "string",
             "title": "Snapshot Pool",
             "format": "x-uri-reference",
-            "default": "/rest/storage-pools/0DB2A6C7-04D3-4830-9229-A80800EEF1F1",
+            "default": "/rest/storage-pools/628C1EBD-5BA7-40F2-A856-A93C0143AC73",
             "description": "A URI reference to the common provisioning group used to create snapshots"
         },
         "provisioningType": {
@@ -128,118 +125,117 @@ options = {
 }
 
 oneview_client = OneViewClient(config)
+storage_pools = oneview_client.storage_pools
+storage_systems = oneview_client.storage_systems
+storage_volume_templates = oneview_client.storage_volume_templates
 
 # Find or add storage pool to use in template
 print("Find or add storage pool to use in template")
-storage_pools = oneview_client.storage_pools.get_all()
+storage_pools_all = storage_pools.get_all()
 storage_pool_added = False
 storage_system_added = False
-if storage_pools:
-    storage_pool = storage_pools[0]
-    print("   Found storage pool '{name}' at uri: '{uri}".format(**storage_pool))
+if storage_pools_all:
+    storage_pool_data = storage_pools_all[0]
+    storage_pool = storage_pools.get_by_uri(storage_pool_data["uri"])
+    print("   Found storage pool '{name}' at uri: '{uri}".format(**storage_pool.data))
 else:
     # Find or add storage system
     storage_pool_added = True
     print("   Find or add storage system")
-    s_systems = oneview_client.storage_systems.get_all()
+    s_systems = storage_systems.get_all()
     if s_systems:
-        s_system = s_systems[0]
+        s_system_data = s_systems[0]
+        storage_system = storage_systems.get_by_uri(s_system_data["uri"])
         storage_system_added = False
-        print("      Found storage system '{name}' at uri: {uri}".format(**s_system))
+        print("      Found storage system '{name}' at uri: {uri}".format(**storage_system.data))
     else:
         options_storage = {
-            "ip_hostname": config['storage_system_hostname'],
+            "hostname": config['storage_system_hostname'],
             "username": config['storage_system_username'],
-            "password": config['storage_system_password']
+            "password": config['storage_system_password'],
+            "family": config['storage_system_family']
         }
-        s_system = oneview_client.storage_systems.add(options_storage)
-        s_system['managedDomain'] = s_system['unmanagedDomains'][0]
-        s_system = oneview_client.storage_systems.update(s_system)
+        storage_system = storage_systems.add(options_storage)
+        s_system_data = storage_system.data.copy()
+        s_system_data['managedDomain'] = storage_system.data['unmanagedDomains'][0]
+        storage_system.update(s_system_data)
         storage_system_added = True
-        print("      Added storage system '{name}' at uri: {uri}".format(**s_system))
+        print("      Added storage system '{name}' at uri: {uri}".format(**storage_system.data))
 
     # Find and add unmanaged storage pool for management
     pool_name = ''
     storage_pool = {}
     print("   Find and add unmanaged storage pool for management")
-    for pool in s_system['unmanagedPools']:
-        if pool['domain'] == s_system['managedDomain']:
+    for pool in storage_system.data['unmanagedPools']:
+        if pool['domain'] == storage_system.data['managedDomain']:
             pool_name = pool['name']
             break
     if pool_name:
         print("      Found pool '{}'".format(pool_name))
         options_pool = {
-            "storageSystemUri": s_system['uri'],
+            "storageSystemUri": storage_system.data['uri'],
             "poolName": pool_name
         }
-        storage_pool = oneview_client.storage_pools.add(options_pool)
+        storage_pool = storage_pools.add(options_pool)
         print("      Successfully added pool")
     else:
         print("      No available unmanaged storage pools to add")
 
 # Create storage volume template
 print("Create storage volume template")
-volume_template = oneview_client.storage_volume_templates.create(options)
-pprint(volume_template)
+volume_template = storage_volume_templates.create(options)
+pprint(volume_template.data)
 
-template_id = volume_template["uri"].split('/')[-1]
+template_id = volume_template.data["uri"].split('/')[-1]
 
 # Update storage volume template
-print("Update '{name}' at uri: {uri}".format(**volume_template))
-volume_template['description'] = "updated description"
-volume_template = oneview_client.storage_volume_templates.update(volume_template)
-print("   Updated with 'description': '{description}'".format(**volume_template))
+print("Update '{name}' at uri: {uri}".format(**volume_template.data))
+volume_template_data = volume_template.data.copy()
+volume_template_data['description'] = "updated description"
+volume_template.update(volume_template_data)
+print("   Updated with 'description': '{description}'".format(**volume_template.data))
 
 # Get all storage volume templates
 print("Get all storage volume templates")
-volume_templates = oneview_client.storage_volume_templates.get_all()
-for template in volume_templates:
+volume_templates_all = storage_volume_templates.get_all()
+for template in volume_templates_all:
     print("   '{name}' at uri: {uri}".format(**template))
 
-# Get storage volume template by id
-try:
-    print("Get storage volume template by id: '{}'".format(template_id))
-    volume_template_byid = oneview_client.storage_volume_templates.get(template_id)
-    print("   Found '{name}' at uri: {uri}".format(**volume_template_byid))
-except HPOneViewException as e:
-    print(e.msg)
-
 # Get storage volume template by uri
-print("Get storage volume template by uri: '{uri}'".format(**volume_template))
-volume_template_by_uri = oneview_client.storage_volume_templates.get(volume_template['uri'])
-print("   Found '{name}' at uri: {uri}".format(**volume_template_by_uri))
+print("Get storage volume template by uri: '{uri}'".format(**volume_template.data))
+volume_template_by_uri = storage_volume_templates.get_by_uri(volume_template.data['uri'])
+print("   Found '{name}' at uri: {uri}".format(**volume_template_by_uri.data))
 
 # Get storage volume template by name
-print("Get storage volume template by 'name': '{name}'".format(**volume_template))
-volume_template_byname = oneview_client.storage_volume_templates.get_by('name', volume_template['name'])[0]
-print("   Found '{name}' at uri: {uri}".format(**volume_template_byname))
+print("Get storage volume template by 'name': '{name}'".format(**volume_template.data))
+volume_template_byname = storage_volume_templates.get_by_name(volume_template.data['name'])
+print("   Found '{name}' at uri: {uri}".format(**volume_template_byname.data))
 
 # Gets the storage templates that are connected on the specified networks
 # scoper_uris and private_allowed_only parameters supported only with API version >= 600
 print("Get torage templates that are connected on the specified networks")
-storage_templates = oneview_client.storage_volume_templates.get_reachable_volume_templates(
+storage_templates = storage_volume_templates.get_reachable_volume_templates(
     networks=networks, scope_uris=scope_uris, private_allowed_only=False)
 print(storage_templates)
 
 # Retrieves all storage systems that is applicable to the storage volume template.
 print("Get storage systems that is applicable to the storage volume template")
-storage_systems = oneview_client.storage_volume_templates.get_compatible_systems(
-    id_or_uri=template_id)
+storage_systems = volume_template.get_compatible_systems()
 print(storage_systems)
 
 # Remove storage volume template
 print("Delete storage volume template")
-oneview_client.storage_volume_templates.delete(volume_template)
+volume_template.delete()
 print("   Done.")
 
 # Remove storage pool
 if storage_pool_added:
     print("Remove recently added storage pool")
-    oneview_client.storage_pools.remove(storage_pool)
+    storage_pool.remove()
     print("   Done.")
 
 # Remove storage system, if it was added
 if storage_system_added:
     print("Remove recently added storage system")
-    oneview_client.storage_systems.remove(s_system)
+    storage_system.remove()
     print("   Done.")